RDKTV-29770: AVOutput plugin code clean-up
Reason for change: code clean-up
Test Procedure: refer ticket
Risks: Low
Priority: P2
